When you are in the "manage questions" tab, you should get a taxonomy dropdown. leave it set at "none" to select random questions from the random question pool you've defined for this quiz. if you set it to a taxonomy term, it will instead choose from questions with that term.
So for example, you could choose the term "fruit", and then any questions that have that are assigned to that term will be eligible to show up on the quiz. Any children of your term will be eligible too. So if "apple" and "orange" were children of fruit, then any questions assigned to those terms could show up on the quiz as well.