On event, check getScore in H5P.instances first
Continously update the 'quiz-h5p-result-key'-field with partial scores,
so that the user can get creadit for correct answers, even if
she continues to the next page without completing all questions.

When an xApi-event is fired, try to use the parent content type to
access the 'getScore', and 'getMaxScore'-functions first. If not
possible, only then check the event itself.

This way we get the cumullative partial score, that can be set in the
'quiz-h5p-result-key'-field in the quiz form.

This commit is also a refactoring to make it easier to reason about
the code.